crate starBurst.js
